+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Fix to store over 256 bytes (up to 1024 bytes)
 ### Added
 * Add multicore_test project
+* Add setDefault() and getDefault() on Parameter<> class
 
 ## [1.0.0] - 2024-06-01
 * Initial release